# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Capell Public Page API
  version: v1
  summary: Versioned public JSON endpoints for published Capell pages and Layout Builder graphs.
  description: Public, read-only page resolution for headless consumers. Returned HTML is sanitized and authoring surfaces are forbidden.
servers:
  - url: https://{host}
    variables:
      host:
        default: example.com
paths:
  /api/capell/v1/pages/resolve:
    get:
      operationId: resolvePublicPageV1
      summary: Resolve a published Capell page by URL.
      description: Canonical v1 endpoint for public page JSON delivery.
      tags:
        - Public Pages
      parameters:
        - $ref: "#/components/parameters/Url"
        - $ref: "#/components/parameters/Site"
        - $ref: "#/components/parameters/Language"
        - $ref: "#/components/parameters/Fields"
        - $ref: "#/components/parameters/Include"
        - $ref: "#/components/parameters/Containers"
      responses:
        "200":
          $ref: "#/components/responses/PageResolved"
        "304":
          $ref: "#/components/responses/NotModified"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/PageNotFound"
        "422":
          $ref: "#/components/responses/InvalidLayoutHtmlRequest"
  /api/capell/pages/resolve:
    get:
      operationId: resolvePublicPageLegacy
      summary: Resolve a published Capell page by URL through the legacy compatibility path.
      description: Compatibility alias for the canonical v1 endpoint. New consumers should use `/api/capell/v1/pages/resolve`.
      tags:
        - Public Pages
      parameters:
        - $ref: "#/components/parameters/Url"
        - $ref: "#/components/parameters/Site"
        - $ref: "#/components/parameters/Language"
        - $ref: "#/components/parameters/Fields"
        - $ref: "#/components/parameters/Include"
        - $ref: "#/components/parameters/Containers"
      responses:
        "200":
          $ref: "#/components/responses/PageResolved"
        "304":
          $ref: "#/components/responses/NotModified"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/PageNotFound"
        "422":
          $ref: "#/components/responses/InvalidLayoutHtmlRequest"
components:
  parameters:
    Url:
      name: url
      in: query
      required: false
      schema:
        type: string
        default: /
        examples:
          - /terms
      description: Published page URL to resolve. Blank or omitted resolves `/`.
    Site:
      name: site
      in: query
      required: false
      schema:
        type: string
      description: Optional signed-context site id. Unsigned requests resolve the site from the request host.
    Language:
      name: language
      in: query
      required: false
      schema:
        type: string
      description: Optional signed-context language id, code, or locale.
    Fields:
      name: fields
      in: query
      required: false
      schema:
        type: string
        enum:
          - url
          - title
          - content
          - meta
          - url,title,content
          - url,title,content,meta
      description: Comma-separated field projection. Unknown fields are ignored.
    Include:
      name: include
      in: query
      required: false
      schema:
        type: string
        enum:
          - layout
          - layout.html
      description: "Optional layout graph include. Rendered widget HTML is only included when `layout.html` is requested."
    Containers:
      name: containers
      in: query
      required: false
      schema:
        type: string
        examples:
          - main
          - main,sidebar
      description: "Comma-separated layout container keys. Required to bound `layout.html` requests."
  responses:
    PageResolved:
      description: Published page resolved.
      headers:
        X-Capell-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Capell-Cache-Tags:
          $ref: "#/components/headers/CacheTags"
        ETag:
          $ref: "#/components/headers/ETag"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PageResponse"
    NotModified:
      description: The current sanitized payload matches the supplied `If-None-Match` validator.
      headers:
        X-Capell-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Capell-Cache-Tags:
          $ref: "#/components/headers/CacheTags"
        ETag:
          $ref: "#/components/headers/ETag"
    Forbidden:
      description: Explicit site or language context was requested without a valid signature.
      headers:
        X-Capell-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Capell-Cache-Tags:
          $ref: "#/components/headers/CacheTags"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              value:
                message: Forbidden
    PageNotFound:
      description: Package, site, language, or page could not be resolved.
      headers:
        X-Capell-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Capell-Cache-Tags:
          $ref: "#/components/headers/CacheTags"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                message: Page not found
    InvalidLayoutHtmlRequest:
      description: "`layout.html` was requested without explicit bounded containers."
      headers:
        X-Capell-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Capell-Cache-Tags:
          $ref: "#/components/headers/CacheTags"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unboundedLayoutHtml:
              value:
                message: layout.html requires explicit bounded containers.
  headers:
    ApiVersion:
      schema:
        type: string
        const: v1
      description: Public response contract version.
    CacheTags:
      schema:
        type: string
      description: Comma-separated cache tags for downstream integrations.
    ETag:
      schema:
        type: string
      description: Strong validator for the selected sanitized JSON payload.
  schemas:
    PageResponse:
      type: object
      required:
        - data
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/PageData"
    PageData:
      type: object
      additionalProperties: false
      properties:
        url:
          type: string
          examples:
            - /terms
        title:
          type:
            - string
            - "null"
        content:
          type:
            - string
            - "null"
          description: Sanitized public HTML.
        meta:
          type:
            - object
            - array
          description: Optional sanitized public page metadata.
        layout:
          $ref: "#/components/schemas/LayoutGraph"
    LayoutGraph:
      type: object
      required:
        - key
        - meta
        - containers
      additionalProperties: false
      properties:
        key:
          type: string
        meta:
          type: object
          additionalProperties: true
        containers:
          type: array
          items:
            $ref: "#/components/schemas/LayoutContainer"
    LayoutContainer:
      type: object
      required:
        - key
        - meta
        - widgets
      additionalProperties: false
      properties:
        key:
          type: string
        meta:
          type: object
          additionalProperties: true
        widgets:
          type: array
          items:
            $ref: "#/components/schemas/LayoutWidget"
    LayoutWidget:
      type: object
      required:
        - key
        - occurrence
        - type
        - data
      additionalProperties: false
      properties:
        key:
          type: string
        occurrence:
          type: integer
          minimum: 1
        type:
          type:
            - string
            - "null"
        data:
          type:
            - object
            - array
            - string
            - number
            - boolean
            - "null"
          description: Sanitized public widget payload.
        component:
          type: string
          description: Optional frontend component identifier when enabled by the host application.
        html:
          type: string
          description: "Optional sanitized rendered widget HTML for bounded `layout.html` requests."
    ErrorResponse:
      type: object
      required:
        - message
      additionalProperties: false
      properties:
        message:
          type: string